Extracted SelectableCell class
